Moved aws to shared library
