+++ backend/aws/__init__.py
@@ -1,0 +1,3 @@
+"""AWS Code Module"""
+
+from aws import AWS
