+++ backend/play_counter/aws
@@ -1,0 +1,1 @@
+../aws
