+++ backend/play_counter/clients/spotify.py
@@ -4,11 +4,10 @@ from datetime import datetime
 from os import environ
 from typing import Union
 
+from aws import AWS
 from models.track import Track
 from spotipy import Spotify, SpotifyOAuth
 from telemetry.logging import logger
-
-from .aws import AWS
 
 
 class SpotifyClient:
